requirements.txt
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.0
frozenlist==1.4.0
googlemaps==4.10.0
h11==0.14.0
idna==3.4
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.1.0
openai==0.28.1
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pydotenv==0.0.7
pymongo==4.5.0
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
Werkzeug==3.0.0
yarl==1.9.2
